// civitas-frontend/src/models/secretaria.ts
export default interface SecretariaDTO {
  idSecretaria: number;
  situacao: number;
  descricao: string;
  cnpj: string;
  nome: string;
  logradouro: string;
  numero: string;
  bairro: string;
  cep: string;
  nomeRazaoSocial: string;
  telefone: string;
  email: string;
  cidade: string;
  estado: string;
}